// apps/admin/src/api/generated/model/assistantThreadListDtoOutput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Join The Six API
 * Operations API for Join The Six
 * OpenAPI spec version: 1.0.0
 */
import type { AssistantThreadListDtoOutputItemsItem } from "./assistantThreadListDtoOutputItemsItem";

export type AssistantThreadListDtoOutput = {
  /** @maxItems 50 */
  items: AssistantThreadListDtoOutputItemsItem[];
};