// apps/admin/src/api/generated/model/createAssistantTurnDto.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Join The Six API
 * Operations API for Join The Six
 * OpenAPI spec version: 1.0.0
 */
import type { CreateAssistantTurnDtoEffort } from "./createAssistantTurnDtoEffort";
import type { CreateAssistantTurnDtoModel } from "./createAssistantTurnDtoModel";

export type CreateAssistantTurnDto = {
  /**
   * @minLength 1
   * @maxLength 20000
   */
  content: string;
  effort?: CreateAssistantTurnDtoEffort;
  model?: CreateAssistantTurnDtoModel;
  /** @pattern ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$ */
  requestId: string;
};